SHOR-42: Restructure files and folders and move gulp architecture to main gulpfile
