--- backstop-config/.gitignore
@@ -1,7 +1,0 @@
-.DS_Store
-node_modules/
-backstop_data/cookies
-backstop_data/screenshots
-backstop_data/*_report
-config/backstop.temp.json
-config/site-config.json
