--- backstop-config/package.json
@@ -1,32 +1,0 @@
-{
-  "name": "backstopjs-Mosaico",
-  "version": "1.0.0",
-  "description": "BackstopJS configuration for CiviCRM extension Mosaico",
-  "main": "index.js",
-  "scripts": {
-    "precommit": "lint-staged",
-    "test": "node test-runner.js"
-  },
-  "lint-staged": {
-    "*.js": [
-      "semistandard --fix",
-      "git add"
-    ]
-  },
-  "devDependencies": {
-    "ansi-colors": "^2.0.2",
-    "backstopjs": "^3.3.1",
-    "clean": "^4.0.2",
-    "fs": "0.0.1-security",
-    "gulp": "^3.9.1",
-    "gulp-clean": "^0.4.0",
-    "gulp-file": "^0.4.0",
-    "gulp-notify": "^3.2.0",
-    "lint-staged": "^6.1.0",
-    "lodash": "^4.17.10",
-    "path": "^0.12.7",
-    "plugin-error": "^1.0.1",
-    "semistandard": "^12.0.0",
-    "yargs": "^4.8.1"
-  }
-}
