--- backstop-config/scenarios/test.json
@@ -1,6 +1,0 @@
-{
-  "scenarios": [{
-    "label": "Dashboard",
-    "url": "{url}/civicrm"
-  }]
-}
