+++ tests/backstop/data/engine_scripts/puppet/clickAndHoverHelper.js
