+++ tests/backstop/data/engine_scripts/puppet/interceptImages.js
